api/requirements.txt
aio-pika==5.6.0
aiodns==2.0.0
aiohttp==3.5.4
aiormq==2.6.0
async-timeout==3.0.1
asyncpg==0.18.3
attrs==19.1.0
cchardet==2.1.4
cffi==1.12.3
chardet==3.0.4
idna==2.8
multidict==4.5.2
pamqp==2.3.0
pycares==3.0.0
pycparser==2.19
PyJWT==1.7.1
yarl==1.3.0
